Day4-6/Day4/day4/pyproject.toml
[tool.poetry]
name = "day4"
version = "0.1.0"
description = ""
authors = ["haiderzeeshan22 <shani.pcr2244@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlmodel = "^0.0.22"
psycopg = {extras = ["binary"], version = "^3.2.3"}
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
